requirements.txt
amqp==2.4.2
billiard==3.5.0.5
celery==4.2.2
Django==2.1.7
django-unixtimestampfield==0.3.9
djangorestframework==3.9.2
Jinja2==2.10
kombu==4.3.0
MarkupSafe==1.1.1
psycopg2-binary==2.7.7
pytz==2018.9
vine==1.3.0
